Add helper to setup WOPI discovery for testing
